fix(interact): document cursor should be restored if target is nullified
The value of `this.target` is defined at `pointerOver` method. But also, at certain points of it, can be set to `null` (`interact.js:1337`). A similar behaviour occurs in `clearTargets` (`interact.js:2633`).

Next time `pointerOut` is called and `this.target` is `null`, the HTML style cursor won't be reseted back to the normal state.

The fix is to call this before the target is nullified:
`this.target._doc.documentElement.style.cursor = '';`